BREAKING CHANGE! Subclasses and package structure
It is time to organize the repo into a better code structure.

* Directory structure structured up.
* What was previously the BulletinMachine is moved into
  SKClassifierMachine, and BulletinMachine is made into an interface.
* .to_dummies() now returns a flat DataFrame.
* Docs removed as they contain the same info as inline comments,
  but are outdated.
* aggregatedata.py split up into itself and download.py
* Dynamic import of varsomdata in one place only in the whole package
  (avaml/__init__.py)
* outputs should go into the output directory
