--- docs/aggregatedata.md
@@ -1,85 +1,0 @@
-# Creating datasets
-`aggregatedata.py` includes the objects `ForecastDataset` and `LabeledData`. Valuable constants
-include `_NONE` and `PROBLEMS`.
-
-## `ForecastDataset`
-This object fetches data from Varsom and RegObs and inserts it into a system of dictionaries
-to facilitate the creation of `LabeledData`. The data is cached, as the response times of the API
-is very slow.
-
-### Constructor: `ForecastDataset(regobs_types: Iterable<Int>, seasons: Iterable<String>)`
-Raises `RegObsRegTypeError` if `regobs_types` is specified incorrectly (if the strings are not a
-subset of the available observations that can be fetched from RegObs).
-
-* `regobs_types`: What to download from RegObs. E.g., if danger signs and snowpack test are
-  wanted, use `("Faretegn", "Tester", "Skredaktivitet")`.
-* `seasons`: The seasons to download and use, e.g. `('2017-18', '2018-19', '2019-20')`.
-
-### Methods
-#### `ForecastDataset.label(self, days: Int) -> LabeledData`
-* `days`: How far back in time values should data be included. 
-  The number of days actually used for each modality is not obvious.
-  The reason for this is to make sure that each kind of data contain
-  the same number of data points, if we want to use some time series
-  frameworks that are picky about such things.
-  * If `0`, only weather data for the forecast day is evaluated.
-  * If `1`, day 0 is used for weather, 1 for Varsom.
-  * If `2`, day 0 is used for weather, 1 for Varsom, 2 for RegObs.
-  * If `3`, days 0-1 is used for weather, 1-2 for Varsom, 2-3 for RegObs.
-  * If `5`, days 0-3 is used for weather, 1-4 for Varsom, 2-5 for RegObs.
-
-## `LabeledData`
-It contains three `DataFrame`s, the features the labels, the predictions and supports some
-operations on those.
-
-### Constructor: `LabeledData(data: DataFrame, label: DataFrame, days: Int, regobs_types: Iterable<String>)`
-* `data`: A `DataFrame` containing features. Probably created by `ForecastDataset.label()`.
-* `label`: A `DataFrame` containing labels. Probably created by `ForecastDataset.label()`.
-* `days`: See `ForecastDataset.label()`.
-* `regobs_types`: See `ForecastDataset()`.
-
-### Methods
-#### `LabeledData.normalize(self) -> LabeledData`
-Scale the data to the range [0, 1].
-
-#### `LabeledData.denormalize(self) -> LabeledData`
-Invert `LabeledData.normalize()`.
-
-#### `LabeledData.kfold(self, k: Int, shuffle: Bool) -> Iterable<(LabeledData, LabeledData)>`
-* `k`: Number of folds.
-* `shuffle`: Whether rows should be shuffled before folding. Defaults to True.
-
-Returns an iterable of LabeledData-tuples. The first element is the training dataset
-and the second is for testing.
-
-#### `LabeledData.f1(self) -> pandas.DataFrame`
-Returns F1, precision, recall of all labels and RMSE of all values.
-
-#### `LabeledData.to_timeseries(self) -> (numpy.ndarray, list<Int>)`
-Converts `LabeledData.data` to a `numpy.ndarray` of shape `(rows, time_steps, modalities)` and a list of
-the names of the features.
-
-#### `LabeledData.to_dummies(self) -> dict<dict<dict<pandas.DataFrame>>>`
-Convert categorical variable into dummy/indicator variables.
-
-The return type should be flattened in the near future (2020-07-06).
-
-#### `LabeledData.to_csv(self)`
-Writes a csv-file in `varsomdata/localstorage` named according to the properties of the dataset.
-A corresponding label file is also always written.
-
-#### `LabeledData.copy(self) -> LabeledData`
-Makes a copy of the object.
-
-### Static methods
-#### `LabeledData.from_csv(days: Int, regobs_types: Iterable<String>) -> LabeledData`
-Reads a csv-file from localstorage with the given properties. Raises a `CsvMissingError` if
-a csv-file with the given properties is not found in localstorage.
-
-* `days`: See `ForecastDataset.label()`.
-* `regobs_types`: See `ForecastDataset()`.
-
-### Attributes
-* `LabeledData.data`: A DataFrame of the features.
-* `LabeledData.label`: A DataFrame of the labels.
-* `LabeledData.pred`: A DataFrame of predicted values.
